+++ src/main/java/dev/shadowsoffire/apotheosis/adventure/AdventureConfig.java
@@ -6,6 +6,7 @@ import dev.shadowsoffire.apotheosis.adventure.loot.LootRarity;
 import dev.shadowsoffire.apotheosis.adventure.loot.RarityClamp;
 import dev.shadowsoffire.apotheosis.adventure.loot.RarityRegistry;
 import dev.shadowsoffire.placebo.config.Configuration;
+import io.github.fabricators_of_create.porting_lib.entity.events.player.PlayerTickEvents;
 import net.minecraft.ResourceLocationException;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceKey;