+++ src/main/java/dev/shadowsoffire/apotheosis/adventure/affix/salvaging/SalvagingMenu.java
@@ -59,6 +59,7 @@ public class SalvagingMenu extends PlaceboContainerMenu {
                 public int getMaxStackSize(ItemStack stack) {
                     return 1;
                 }
+
             });
         }