+++ src/main/java/dev/shadowsoffire/apotheosis/mixin/adventure/BowItemMixin.java
@@ -17,8 +17,11 @@ public class BowItemMixin {
 
     @WrapWithCondition(method = "releaseUsing", at = @At(value = "INVOKE",target = "net/minecraft/world/level/Level.addFreshEntity (Lnet/minecraft/world/entity/Entity;)Z"))
     private boolean arrowEvent(Level instance, Entity entity) {
-        if (Apotheosis.enableDebug) AdventureModule.LOGGER.info("Current arrow base damage pre event: {}", ((AbstractArrow) entity).getBaseDamage());
-        AdventureEvents.fireArrow((AbstractArrow) entity);
+        if (Apotheosis.enableAdventure) {
+            if (Apotheosis.enableDebug)
+                AdventureModule.LOGGER.info("Current arrow base damage pre event: {}", ((AbstractArrow) entity).getBaseDamage());
+            AdventureEvents.fireArrow((AbstractArrow) entity);
+        }
         return true;
     }
 }