+++ src/main/java/dev/shadowsoffire/apotheosis/mixin/adventure/EnchantmentHelperMixinAdventure.java
@@ -1,5 +1,6 @@
 package dev.shadowsoffire.apotheosis.mixin.adventure;
 
+import dev.shadowsoffire.apotheosis.Apotheosis;
 import dev.shadowsoffire.apotheosis.adventure.affix.AffixHelper;
 import dev.shadowsoffire.apotheosis.adventure.affix.AffixInstance;
 import net.minecraft.world.damagesource.DamageSource;
@@ -21,7 +22,8 @@ public class EnchantmentHelperMixinAdventure {
      * Injection to {@link EnchantmentHelper#getDamageProtection(Iterable, DamageSource)}
      */
     @Inject(at = @At("RETURN"), method = "getDamageProtection(Ljava/lang/Iterable;Lnet/minecraft/world/damagesource/DamageSource;)I", cancellable = true)
-    private static void apoth_getDamageProtection(Iterable<ItemStack> stacks, DamageSource source, CallbackInfoReturnable<Integer> cir) {
+    private static void zenith_getDamageProtection(Iterable<ItemStack> stacks, DamageSource source, CallbackInfoReturnable<Integer> cir) {
+        if (!Apotheosis.enableAdventure) return;
         int prot = cir.getReturnValueI();
         for (ItemStack s : stacks) {
             var affixes = AffixHelper.getAffixes(s);
@@ -36,7 +38,8 @@ public class EnchantmentHelperMixinAdventure {
      * Injection to {@link EnchantmentHelper#getDamageBonus(ItemStack, MobType)
      */
     @Inject(at = @At("RETURN"), method = "getDamageBonus(Lnet/minecraft/world/item/ItemStack;Lnet/minecraft/world/entity/MobType;)F", cancellable = true)
-    private static void apoth_getDamageBonus(ItemStack stack, MobType type, CallbackInfoReturnable<Float> cir) {
+    private static void zenith_getDamageBonus(ItemStack stack, MobType type, CallbackInfoReturnable<Float> cir) {
+        if (!Apotheosis.enableAdventure) return;
         float dmg = cir.getReturnValueF();
         var affixes = AffixHelper.getAffixes(stack);
         for (AffixInstance inst : affixes.values()) {
@@ -50,6 +53,7 @@ public class EnchantmentHelperMixinAdventure {
      */
     @Inject(at = @At("TAIL"), method = "doPostDamageEffects(Lnet/minecraft/world/entity/LivingEntity;Lnet/minecraft/world/entity/Entity;)V")
     private static void zenith_doPostDamageEffects(LivingEntity user, Entity target, CallbackInfo ci) {
+        if (!Apotheosis.enableAdventure) return;
         if (user == null) return;
         for (ItemStack s : user.getAllSlots()) {
             var affixes = AffixHelper.getAffixes(s);
@@ -67,6 +71,7 @@ public class EnchantmentHelperMixinAdventure {
      */
     @Inject(at = @At("TAIL"), method = "doPostHurtEffects(Lnet/minecraft/world/entity/LivingEntity;Lnet/minecraft/world/entity/Entity;)V")
     private static void zenith_doPostHurtEffects(LivingEntity user, Entity attacker, CallbackInfo ci) {
+        if (!Apotheosis.enableAdventure) return;
         if (user == null) return;
         for (ItemStack s : user.getAllSlots()) {
             var affixes = AffixHelper.getAffixes(s);